Update README.md
On the IDT, I do not clearly see the idea that the team can just disclose "all tokens are held in a specific wallet address." Is this in a particular subparagraph? If so, I think using plain language it should say exactly that. And that any transfers from that address can thereafter be tracked.

Re: Notice of reliance, maybe make it an AND/OR type of statement about stamping directly into reliance. Probably easier to maintain if stamped directly into the token as opposed to maintaining a separate website as a version of custodian of records. 

Nevertheless, for accessibility and ease of use purposes, there may be occasions for why it should be hosted in on a website in addition to being stamped directly on the token.

Line 88: Final sentence really needs to be re-worked for clarity: "(v)  To the extent the network allows it, this notice of reliance may be provided in the source code of the token itself as a boolean or string statement of reliance that immutably links to website or IPFS hash that provides this disclosure.  This reliance written in the Token would provide notice to all parties using the blockchain reliance is being made by the Intitial Development Team on this Safe Harbor."